--- app/Ideogram/categories.ts
@@ -1,128 +1,0 @@
-export default [
-  {
-    name: 'A Locus Agouti Signaling Protein (ASIP)',
-    location: 'chr24:23,334,340-23,373,371',
-    type: 'coatcolor',
-    summary:
-      "The Agouti Signaling Protein (ASIP) gene plays a crucial role in determining coat color patterns in dogs. It regulates the production of red and yellow pigments (pheomelanin) versus black and brown pigments (eumelanin) in hair follicles. In dogs, different variants of the ASIP gene are associated with various coat color patterns, including sable, fawn, and black-and-tan. The gene's activity can be modified by other genes, leading to a wide range of coat color variations across different dog breeds. Mutations in the ASIP gene can result in unique coat patterns, such as the dominant black coloration seen in some breeds like German Shepherds and Shetland Sheepdogs. Further reading: (Bannasch et al., 2021) https://www.nature.com/articles/s41559-021-01524-x",
-  },
-  {
-    name: 'E Locus Melanocortin 1 receptor (MC1R)',
-    location: 'chr5:63,942,798-63,943,751',
-    type: 'coatcolor',
-    summary:
-      'The Melanocortin 1 receptor (MC1R) gene is a key determinant of pigmentation in dogs. It controls the production of eumelanin (black/brown pigment) versus pheomelanin (red/yellow pigment) in melanocytes. Variations in the MC1R gene are responsible for different coat colors in dogs, including red, yellow, and cream. Mutations in this gene can lead to a loss of function, resulting in the production of only red/yellow pigment, as seen in Irish Setters and yellow Labrador Retrievers. The interaction between MC1R and other coat color genes, such as ASIP, contributes to the wide variety of coat colors and patterns observed in different dog breeds.',
-  },
-  {
-    name: 'K Locus β-defensin 103 (CBD103)',
-    location: 'chr16:59,123,440-59,124,995',
-    type: 'coatcolor',
-    summary:
-      'The β-defensin 103 (CBD103) gene, also known as the K locus, plays a significant role in determining coat color in dogs. It produces a protein that can bind to the Melanocortin 1 receptor (MC1R), influencing pigment production. The dominant "black" allele of this gene (KB) causes solid black coloration in many dog breeds, overriding other color patterns. The CBD103 gene interacts with other coat color genes, particularly ASIP, to produce various color patterns. Mutations in this gene are responsible for the brindle coat pattern seen in breeds like Boxers and Great Danes, as well as the dominant black coloration in breeds such as Labrador Retrievers and German Shepherds. Further reading: (Candille et al., 2007) https://www.science.org/doi/10.1126/science.1147880 (Nordlund et al., 2006) https://www.google.com/books/edition/The_Pigmentary_System/4sn46U1uexoC?hl=en&gbpv=1&pg=PR5&printsec=frontcover',
-  },
-  {
-    name: 'D Locus Dilute/Blue melanophilin (MLPH)',
-    location: 'chr25:48,382,738-48,428,452',
-    type: 'coatcolor',
-    summary:
-      'The melanophilin (MLPH) gene, also known as the D locus, is responsible for the dilution of coat colors in dogs. It plays a crucial role in the transport and distribution of pigment granules within melanocytes. Mutations in the MLPH gene can lead to improper pigment distribution, resulting in a diluted or "washed-out" appearance of the coat color. This dilution effect transforms black to blue/grey, liver to isabella (lilac), and red to cream. The dilute phenotype is observed in various dog breeds, including Weimaraners, Doberman Pinschers, and some Great Danes. In some breeds, the dilute phenotype has been associated with coat problems such as color dilution alopecia, although this link is not fully understood and may involve other genetic factors. Further reading: (Van Buren, 2020) https://www.mdpi.com/2073-4425/11/6/639',
-  },
-  {
-    name: 'M Locus Merle premelanosome protein (PMEL17)',
-    location: 'chr10:277,298-278,441',
-    type: 'coatcolor',
-    summary:
-      'The premelanosome protein (PMEL17) gene, also known as the M locus or SILV gene, is responsible for the merle coat pattern in dogs. It plays a crucial role in the formation of melanosomes, the organelles where melanin pigments are produced and stored. A specific insertion in the PMEL17 gene causes the merle phenotype, characterized by patches of diluted pigment interspersed with normal coloration. The merle pattern is seen in various dog breeds, including Australian Shepherds, Collies, and Great Danes. However, the merle mutation can have health implications when inherited in a double dose (homozygous merle), often resulting in vision and hearing problems. Due to these potential health issues, breeding two merle dogs together is generally discouraged in responsible breeding practices.',
-  },
-  {
-    name: 'H Locus Harlequin proteasome 20S subunit beta 7 (PSMB7)',
-    location: 'chr9:58,207,216-58,270,483',
-    type: 'coatcolor',
-    summary:
-      'The proteasome 20S subunit beta 7 (PSMB7) gene, also known as the H locus, is responsible for the Harlequin coat pattern in Great Danes. This gene interacts with the merle gene (PMEL17) to produce the distinctive Harlequin pattern, characterized by white dogs with black patches. The Harlequin mutation in PSMB7 modifies the merle phenotype by removing the dilute patches, leaving only the full-color patches on a white background. This pattern is unique to Great Danes and is not found in other dog breeds. However, like the merle gene, the Harlequin gene can have health implications when inherited in certain combinations, particularly when two Harlequin dogs are bred together. Responsible breeding practices typically avoid breeding Harlequin to Harlequin to prevent potential health issues in offspring.',
-  },
-  {
-    name: 'B Locus Brown/Chocolate/Liver tyrosinase related protein 1 (TYRP1)',
-    location: 'chr11:32,928,733-32,946,547',
-    type: 'coatcolor',
-    summary:
-      "The tyrosinase related protein 1 (TYRP1) gene, also known as the B locus, is responsible for brown coat color in dogs. Great Danes typically do not have the recessive B locus allele (b), which is responsible for brown coats in most other breeds. However, there is a small chance that Great Danes could have the recessive b allele. It plays a crucial role in melanin synthesis, specifically in the production of eumelanin (black and brown pigments). Mutations in the TYRP1 gene can result in the production of brown pigment instead of black, leading to coat colors often referred to as brown, chocolate, or liver depending on the breed. This gene affects both coat color and the color of the dog's nose, paw pads, and eyes. The brown color produced by TYRP1 mutations can be seen in many dog breeds, including Labrador Retrievers, German Shorthaired Pointers, and Chesapeake Bay Retrievers. Interestingly, the intensity of the brown color can vary depending on other genes and modifiers, leading to a range of shades from light to dark brown.",
-  },
-  {
-    name: 'S Locus Microphthalmia Associated Transcription Factor (MITF)',
-    location: 'chr20:21,572,611-21,603,715',
-    type: 'coatcolor',
-    summary:
-      'The Microphthalmia Associated Transcription Factor (MITF) gene, also known as the S locus, is a key regulator of melanocyte development and function in dogs. It plays a crucial role in determining the presence and distribution of pigment-producing cells in the skin and hair follicles. Variations in the MITF gene are associated with white spotting patterns in dogs, ranging from small white markings to largely white coats with colored patches. The extent of white spotting can vary widely, from the minimal white chest spot seen in some breeds to the predominantly white coats with colored head and body spots seen in others. MITF mutations can also affect eye color and hearing in some dogs, particularly in breeds with extensive white coloration. This gene is responsible for piebald and extreme white spotting patterns seen in many breeds, including Beagles, Bull Terriers, and Dalmatians.',
-  },
-  {
-    name: 'Serpin family F member 2 (SERPINF2)',
-    location: 'chr9:45,474,549-45,480,834',
-    type: 'health',
-    summary:
-      'The SERPINF2 gene, which stands for Serpin Family F Member 2, encodes a protein called alpha-2-antiplasmin, a key regulator of blood clotting in mammals, including dogs. In canines, this gene plays a crucial role in maintaining proper hemostasis by inhibiting plasmin, an enzyme that breaks down blood clots. Mutations in the SERPINF2 gene have been associated with a rare bleeding disorder in dogs called alpha-2-antiplasmin deficiency, which can lead to prolonged bleeding times and increased risk of hemorrhage. While this condition is uncommon, it has been reported in various dog breeds and can be a concern for breeders and veterinarians managing canine health. Understanding the function of SERPINF2 in dogs not only contributes to better diagnostic and treatment options for affected animals but also provides valuable insights into comparative studies of blood coagulation disorders across species, including humans. Further reading: (Court et al., 2023) https://onlinelibrary.wiley.com/doi/full/10.1111/jvim.16643',
-  },
-  {
-    name: 'Superoxide dismutase 1 (SOD1)',
-    location: 'chr31:26,600,785-26,608,887',
-    type: 'health',
-    summary:
-      'The Superoxide dismutase 1 (SOD1) gene plays a crucial role in protecting cells from oxidative stress by breaking down potentially harmful oxygen molecules. In dogs, mutations in the SOD1 gene have been associated with degenerative myelopathy (DM), a progressive neurological disorder that affects the spinal cord. This condition is particularly prevalent in certain breeds, including German Shepherds, Pembroke Welsh Corgis, and Boxers. DM typically manifests in older dogs and leads to gradual paralysis of the hind limbs, eventually progressing to affect the front limbs. While not all dogs with SOD1 mutations will develop DM, genetic testing for these mutations can help breeders make informed decisions to reduce the incidence of the disease in susceptible breeds. Understanding the role of SOD1 in canine health has significant implications for both veterinary medicine and comparative studies of human amyotrophic lateral sclerosis (ALS), as similar SOD1 mutations are associated with some forms of ALS in humans.',
-  },
-  {
-    name: 'Bridging integrator 1 (BIN1)',
-    location: 'chr19:23,511,496-23,564,133',
-    type: 'health',
-    summary:
-      'The Bridging integrator 1 (BIN1) gene plays a crucial role in muscle function and has been associated with centronuclear myopathy (CNM) in dogs. CNM is a severe muscle disorder that causes muscle weakness and atrophy, particularly affecting Great Danes. Mutations in the BIN1 gene can lead to improper muscle fiber formation and function, resulting in affected dogs having difficulty standing, walking, and even swallowing. The condition is inherited in an autosomal recessive manner, meaning that dogs must inherit two copies of the mutated gene (one from each parent) to develop the disease. Genetic testing for BIN1 mutations has become an important tool for Great Dane breeders to reduce the incidence of CNM in the breed. Research on BIN1 and CNM in dogs has also provided valuable insights into similar muscle disorders in humans, highlighting the importance of canine genetic studies in comparative medicine.',
-  },
-  {
-    name: 'Major Facilitator Superfamily Domain Containing 12 gene (MFSD12)',
-    location: 'chr20:55,570,178-55,577,999',
-    type: 'health',
-    summary:
-      'The Major Facilitator Superfamily Domain Containing 12 (MFSD12) gene has been identified as a key player in pigmentation, particularly in the production of gray coat color in dogs. This gene is involved in the transport of molecules across cellular membranes and affects the type of pigment produced in melanocytes. Mutations in MFSD12 have been associated with the production of gray or "blue" coat colors in several dog breeds, including Kerry Blue Terriers and Italian Greyhounds. The gene\'s effects are particularly noticeable in breeds where the gray color develops over time, with puppies born dark and gradually lightening as they mature. While primarily known for its role in coat color, research into MFSD12 also provides insights into the complex genetic interactions that determine canine pigmentation patterns. Understanding the function of MFSD12 contributes to our broader knowledge of melanin production and distribution in mammals, including humans.',
-  },
-  {
-    name: 'N-acyl phosphatidylethanolamine phospholipase D (NAPEPLD)',
-    location: 'chr18:17,032,965-17,062,294',
-    type: 'health',
-    summary:
-      'The N-acyl phosphatidylethanolamine phospholipase D (NAPEPLD) gene has been associated with a condition known as Hereditary Nasal Parakeratosis (HNPK) in dogs. HNPK is characterized by the excessive production of keratin in the nasal planum (the hairless part of the nose), leading to a dry, crusted, and cracked appearance of the nose. This condition has been particularly noted in Labrador Retrievers, but can occur in other breeds as well. The NAPEPLD gene is involved in the metabolism of certain lipids that are important for maintaining the barrier function of the skin, including the specialized skin of the nose. Mutations in this gene can disrupt the normal balance of lipids in the nasal epithelium, leading to the characteristic symptoms of HNPK. Genetic testing for NAPEPLD mutations allows breeders to identify carriers and affected dogs, helping to reduce the incidence of this condition in susceptible breeds.',
-  },
-  {
-    name: 'Progressive rod-cone degeneration (PRCD)',
-    location: 'chr9:4,886,683-4,889,985',
-    type: 'health',
-    summary:
-      "The Progressive rod-cone degeneration (PRCD) gene is associated with a form of progressive retinal atrophy (PRA) in dogs, a group of inherited eye disorders that lead to blindness. PRCD-PRA is one of the most common forms of PRA, affecting numerous dog breeds including Labrador Retrievers, Poodles, and Cocker Spaniels. The PRCD gene plays a crucial role in the maintenance and function of photoreceptor cells in the retina. Mutations in this gene lead to the gradual degeneration of these cells, first affecting night vision (rod cells) and later day vision (cone cells), ultimately resulting in complete blindness. The condition is inherited in an autosomal recessive manner, meaning a dog must inherit two copies of the mutated gene to develop the disease. Genetic testing for PRCD mutations has become a valuable tool for breeders to reduce the incidence of this form of PRA in affected breeds, and for owners to plan for their dog's potential vision loss.",
-  },
-  {
-    name: 'General transcription factor II-I (GTF2I)',
-    location: 'chr6:5,468,118-5,648,755',
-    type: 'hypersocial',
-    summary:
-      'The General transcription factor II-I (GTF2I) gene has gained attention in canine genetics for its potential role in social behavior. This gene is part of a region that has been associated with Williams-Beuren syndrome in humans, a condition characterized by hypersociability among other traits. In dogs, variations in the GTF2I gene have been linked to hypersocial behavior, particularly the tendency to seek attention and interaction from humans. Studies have found that domestic dogs tend to have variants of this gene that may contribute to their characteristic friendliness towards humans, in contrast to wolves. The GTF2I gene is thought to play a role in regulating other genes involved in social behavior and cognitive development. Research into this gene provides insights into the genetic basis of the unique bond between dogs and humans, and contributes to our understanding of the domestication process.',
-  },
-  {
-    name: 'General transcription factor II-I repeat domain-containing protein 1 (GTF2IRD1)',
-    location: 'chr6:5,695,854-5,806,538',
-    type: 'hypersocial',
-    summary:
-      'The General transcription factor II-I repeat domain-containing protein 1 (GTF2IRD1) gene is another gene associated with social behavior in dogs, often studied alongside GTF2I. Like GTF2I, GTF2IRD1 is part of the region linked to Williams-Beuren syndrome in humans, which is characterized by hypersocial behavior among other traits. In dogs, variations in GTF2IRD1 have been associated with differences in social behavior, particularly in how dogs interact with humans. This gene is thought to play a role in regulating the expression of other genes involved in neural development and function. Studies have suggested that changes in GTF2IRD1 may have contributed to the evolution of dog behavior during domestication, potentially influencing traits that made dogs more suited to living and working with humans. Research into GTF2IRD1 and related genes provides valuable insights into the genetic basis of dog-human bonding and the complex interplay between genetics and behavior in canines.',
-  },
-  {
-    name: 'Putative polypeptide N-acetylgalactosaminyltransferase-like protein 3 (WBSCR17)',
-    location: 'chr6:1,873,204-2,303,372',
-    type: 'hypersocial',
-    summary:
-      'The Putative polypeptide N-acetylgalactosaminyltransferase-like protein 3 (WBSCR17) gene, also known as GALNT17, is another gene located in the region associated with Williams-Beuren syndrome in humans. In dogs, this gene has been studied for its potential role in social behavior and cognitive development. While its specific function in dogs is not as well understood as GTF2I or GTF2IRD1, variations in WBSCR17 have been associated with differences in social behavior and cognitive traits across dog breeds. The gene is thought to be involved in protein modification processes that may influence neural development and function. Research into WBSCR17 and other genes in this region contributes to our understanding of the genetic basis of complex behavioral traits in dogs. Studying these genes helps shed light on the evolutionary changes that occurred during dog domestication and the genetic underpinnings of the unique cognitive and social abilities of domestic dogs.',
-  },
-  {
-    name: 'Nuclear envelope pore membrane protein (POM121)',
-    location: 'chr6:6,688,239-6,715,065',
-    type: 'hypersocial',
-    summary:
-      "The Nuclear envelope pore membrane protein (POM121) gene plays a crucial role in the formation and function of nuclear pore complexes, which are essential for the transport of molecules between the cell nucleus and cytoplasm. While POM121 is not as well-studied in dogs as some of the other genes we've discussed, it is known to be important for cellular function across many species. In the context of canine genetics, variations in POM121 could potentially influence cellular processes that affect development, growth, or various physiological functions. Some studies have suggested that genes involved in nuclear pore complex formation, including POM121, may play a role in cellular stress responses and longevity. However, more research is needed to fully understand the specific implications of POM121 variations in dogs. Studying genes like POM121 contributes to our broader understanding of canine cellular biology and may provide insights into developmental processes or potential health implications in dogs.",
-  },
-] as const
